分类、推荐、搜索
新增搜索界面
修复分类、推荐界面在android4.1上的问题
新建趣处详情界面、图片资源

+++ QuChu/app/src/main/java/co/quchu/quchu/model/ClassifyModel.java
@@ -19,7 +19,7 @@ public class ClassifyModel {
     private String en;
     private String maxImg;
     private String minImg;
-    private int weight;
+    private double weight;
     private String zh;
 
     public void setEn(String en) {
@@ -34,7 +34,7 @@ public class ClassifyModel {
         this.minImg = minImg;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
 
@@ -54,7 +54,7 @@ public class ClassifyModel {
         return minImg;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
